backend/requirements.txt
#flask
psycopg2
uwsgi
Flask
Flask-Migrate
Flask-SQLAlchemy
